Move resolution to own module
